php7.3のユニットテストが失敗していたので調整

+++ docker/apache/init.sh
@@ -1,12 +1,20 @@
 if [ ! -e '/var/www/shared/docker/check' ]; then
-    rm -rf /var/www/shared/tmp
-    rm -rf /var/www/shared/logs
-	rm /var/www/shared/config/jwt.key
-	rm /var/www/shared/config/jwt.pem
-    rsync -av /var/www/shared/ /var/www/html --exclude='node_modules' --exclude='tmp' --exclude='.git' --exclude='.idea' --exclude='.DS_Store' --exclude='docker'
+    if [ -d '/var/www/shared/tmp' ]; then
+        rm -rf /var/www/shared/tmp
+    fi
+    if [ -d '/var/www/shared/logs' ]; then
+        rm -rf /var/www/shared/logs
+    fi
+    if [ -e '/var/www/shared/config/jwt.key' ]; then
+	    rm /var/www/shared/config/jwt.key
+	fi
+	if [ -e '/var/www/shared/config/jwt.pem' ]; then
+	    rm /var/www/shared/config/jwt.pem
+	fi
+    rsync -a /var/www/shared/ /var/www/html --exclude='node_modules' --exclude='tmp' --exclude='.git' --exclude='.idea' --exclude='.DS_Store' --exclude='docker'
     composer install --no-plugins
     cp /var/www/html/config/.env.example /var/www/html/config/.env
-    sleep 10
+    sleep 5
 	/var/www/html/bin/cake migrations migrate --plugin BaserCore
 	/var/www/html/bin/cake migrations seed --plugin BaserCore
 	/var/www/html/bin/cake plugin assets symlink
